requirements.txt
httpx==0.23.3
python-telegram-bot==13.13
openai==0.27.4